fix/edit-title bug and add api
