+++ app/javascript/editor/index.js
@@ -1,6 +1,6 @@
 // import your editor js here
 import './new'
-import './save_button'
+import './edit_title'
 import './delete_button'
 import './restore_or_del_button'
 import './update_button'
