+++ app/javascript/frontend/index.js
@@ -1,3 +1,4 @@
 import './fontawesome'
 import './frontend'
+import './pen'
 import './modal'
