+++ app/javascript/frontend/pen.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+ReactDOM.render(
+  <h1>Hello, world!!</h1>,
+  document.getElementById('cccc')
+);
