+++ app/views/pens/_pens_grid.html.erb
@@ -1,3 +1,4 @@
+<div id="cccc"></div>
 <div class="pen-items-wrap">
   <%# 測試隔線 入資料自行刪除,留一個article %>
   <article class="col">
